requirements.txt
streamlit
langchain>=0.2.0
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
pypdf
python-dotenv
